+++ src/screens/App/screens/SearchBar/index.js
