+++ src/screens/App/screens/SearchBar/screens/FilterSelector/strings/strings.js
