+++ src/screens/App/screens/SearchBar/screens/FilterSelector/style.css
