+++ src/screens/App/screens/SearchBar/screens/SearchComponent/index.js
