+++ src/screens/App/screens/SearchBar/screens/SearchComponent/style.css
