+++ src/screens/App/screens/SearchBar/style.css
